CHORE: move deregister to recruit.py
